server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["James Bristow <jbris@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
plotly = "^5.19.0"
pandas = "^2.2.1"
scipy = "^1.12.0"
scikit-learn = "^1.4.1.post1"
matplotlib = "^3.8.3"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
strawberry-graphql = {extras = ["debug-server"], version = "^0.219.2"}
joblib = "^1.3.2"
graphql-core = "^3.2.3"
sqlalchemy = "^2.0.27"
pydantic = "^2.6.2"
redis = "^5.0.1"
elasticsearch = "^8.12.1"
bokeh = "^3.3.4"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
